requirements.txt
autopep8==1.5.3
backcall==0.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
mccabe==0.6.1
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tweepy==3.9.0
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
